pyproject.toml
[tool.poetry]
name = "freshdesk-token-exchange"
version = "1.0.0"
description = ""
authors = ["30ème Ciel <admin@30emeciel.fr>"]

[tool.poetry.dependencies]
python = "^3.8"
dotmap = "^1.3.22"
flask = "^1.1.2"
PyJWT = "^2.0.1"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
